Add navigation to exit point

+++ ORB_SLAM2/Examples/Monocular/point.cpp
@@ -1,24 +1,20 @@
-//
-// Created by saker on 9/7/2021.
-//
-
 #include "point.hpp"
 
-Point::Point(double x, double y) : x(x),y(y) {}
+Point::Point(float x, float y) : x(x),y(y) {}
 
-double Point::getX() const {
+float Point::getX() const {
     return x;
 }
 
-double Point::getY() const {
+float Point::getY() const {
     return y;
 }
 
-void Point::setX(double x) {
+void Point::setX(float x) {
     Point::x = x;
 }
 
-void Point::setY(double y) {
+void Point::setY(float y) {
     Point::y = y;
 }
 
@@ -32,7 +28,7 @@ void Point::setY(std::string str) {
 
 double convert (std::string str){
     int i,j;
-    double num = 0, q = 0;
+    float num = 0, q = 0;
     bool Int = true, sign = false;
 
     for (i = 0; i < str.length(); i++) {
